chore: rename file update.rs to change.rs to reflect its purpose
