+++ src/seq_value/mod.rs
@@ -18,13 +18,13 @@
 //! sequence number and optional metadata. These are used to track the ordering and
 //! lifecycle of values in storage systems.
 
+mod change;
 mod seq_value_trait;
 mod seqv;
-mod update;
 
+pub use change::Change;
 pub use seq_value_trait::SeqValue;
 pub use seqv::SeqV;
-pub use update::Change;
 
 #[deprecated(since = "0.2.0", note = "Use `Change` instead")]
 pub type Update<M, V = Vec<u8>> = Change<M, V>;
